+++ app/car-marketplace/page.tsx
@@ -1,0 +1,5 @@
+import CarMarketplaceClient from "./CarMarketplaceClient";
+
+export default function CarMarketplacePage() {
+  return <CarMarketplaceClient />;
+}