+++ app/liked-cars-page/page.tsx
@@ -1,0 +1,5 @@
+import LikedCarsPageClient from "./LikedCarsPageClient";
+
+export default function LikedCarsPage() {
+  return <LikedCarsPageClient />;
+}