+++ app/liked-cars/page.tsx
@@ -1,0 +1,5 @@
+import LikedCarsClient from "./LikedCarsClient";
+
+export default function LikedCarsPage() {
+  return <LikedCarsClient />;
+}