+++ app/location/page.tsx
@@ -1,0 +1,5 @@
+import LocationClient from "./LocationClient";
+
+export default function LocationPageRoute() {
+  return <LocationClient />;
+}