+++ app/upload-vehicle/page.tsx
@@ -1,0 +1,5 @@
+import UploadVehicleClient from "./UploadVehicleClient";
+
+export default function UploadVehiclePage() {
+  return <UploadVehicleClient />;
+}